### 2019-12-08 #1939 @danielskatz
👋 @oxinabox  - thanks for your submission!

As you can see above, @whedon thinks some of your bib entries might be missing DOIs (but @whedon is also sometimes wrong)

Please check

There also seem to be some other things in the generated pdf that could be fixed, such as double (())s around some citations, missing spaces before some citations, and some extra "s in one reference

After you make changes, enter a new comment here with ```@whedon generate pdf``` to check the result
