### 2019-12-09 571060d
Add DOIs to paper
