+++ paper/paper.bib
@@ -7,7 +7,8 @@
   Eprinttype               = {arXiv},
   Keywords                 = {tools},
   Timestamp                = {2015.12.18},
-  Url                      = {http://arxiv.org/abs/1411.1607}
+  Url                      = {http://arxiv.org/abs/1411.1607},
+  Doi                      = {10.1137/141000671}
 }
 
 @Inproceedings{NLTK1,
@@ -64,7 +65,8 @@
   eprint    = {1804.08205},
   timestamp = {Mon, 13 Aug 2018 16:49:01 +0200},
   biburl    = {https://dblp.org/rec/bib/journals/corr/abs-1804-08205},
-  bibsource = {dblp computer science bibliography, https://dblp.org}
+  bibsource = {dblp computer science bibliography, https://dblp.org},
+  doi       = {10.1609/aaai.v33i01.33016843}
 }
 
 @online{reversibletok2,
