### 2019-12-09 7c44ada
Fix double brackets and extra "s in the paper
