+++ paper/paper.bib
@@ -24,7 +24,7 @@
 @Book{NLTK2,
   Title                    = {Natural language processing with Python},
   Author                   = {Bird, Steven and Klein, Ewan and Loper, Edward},
-  Publisher                = {" O'Reilly Media, Inc."},
+  Publisher                = {O'Reilly Media, Inc.},
   Year                     = {2009},
   Keywords                 = {software, tools},
   Timestamp                = {2015.07.12},
