### 2019-12-13 #1956 @whedon
[REVIEW]: WordTokenizers.jl: Basic tools for tokenizing natural language in Julia
**Submitting author:** @oxinabox (<a href="http://orcid.org/0000-0003-1386-1646">Lyndon White</a>)
**Repository:** <a href="https://github.com/JuliaText/WordTokenizers.jl/" target ="_blank">https://github.com/JuliaText/WordTokenizers.jl/</a>
**Version:** v0.5.4
