### 2020-01-18 5d4662a
fixing a number of typos in paper and readme
