+++ README.md
@@ -7,8 +7,8 @@
 
 Some basic tokenizers for Natural Language Processing:
 
-The normal way to used this package is to call
-`tokenize(str)` to split up a string into words;
+The normal way to use this package is to call
+`tokenize(str)` to split up a string into words
 or `split_sentences(str)` to split up a string into sentences.
 Maybe even `tokenize.(split_sentences(str))` to do both.
 
@@ -17,7 +17,7 @@ that call one of the tokenizers or sentence splitters defined below.
 They have sensible defaults set,
 but you can override the method used by calling
 `set_tokenizer(func)` or `set_sentence_splitter(func)` passing in your preferred
-function `func` from the list below (or from else where)
+function `func` from the list below (or from elsewhere)
 Configuring them this way will throw up a method overwritten warning, and trigger recompilation of any methods that use them.
 
 This means if you are using a package that uses WordTokenizers.jl to do tokenization/sentence splitting via the default methods,
@@ -30,7 +30,7 @@ If as a package author you don't want to allow the user to change the tokenizer 
 
 ### Example Setting Tokenizer  (Revtok.jl)
 You might like to, for example use [Revtok.jl's tokenizer](https://github.com/jekbradbury/Revtok.jl).
-We do not include Revtok in this package, because making use of it within WordTokenizers.jl is trival.
+We do not include Revtok in this package, because making use of it within WordTokenizers.jl is trivial.
 Just `import Revtok; set_tokenizer(Revtok.tokenize)`.
 
 
@@ -63,20 +63,20 @@ The word tokenizers basically assume sentence splitting has already been done.
  - **Poorman's tokenizer:** (`poormans_tokenize`) Deletes all punctuation, and splits on spaces. (In some ways worse than just using `split`)
  - **Punctuation space tokenize:** (`punctuation_space_tokenize`) Marginally improved version of the poorman's tokenizer, only deletes punctuation occurring outside words.
 
- - **Penn Tokenizer:** (`penn_tokenize`) This is Robert MacIntyre's orginal tokenizer used for the Penn Treebank. Splits contractions.
+ - **Penn Tokenizer:** (`penn_tokenize`) This is Robert MacIntyre's original tokenizer used for the Penn Treebank. Splits contractions.
  - **Improved Penn Tokenizer:** (`improved_penn_tokenize`) NLTK's improved Penn Treebank Tokenizer. Very similar to the original, some improvements on punctuation and contractions. This matches to NLTK's `nltk.tokenize.TreeBankWordTokenizer.tokenize`.
- - **NLTK Word tokenizer:** (`nltk_word_tokenize`) NLTK's even more improved version of the Penn Tokenizer. This version has better unicode handling and some other changes. This matches to the most commonly used `nltk.word_tokenize`, minus the sentence tokenizing step.
+ - **NLTK Word tokenizer:** (`nltk_word_tokenize`) NLTK's even more improved version of the Penn Tokenizer. This version has better Unicode handling and some other changes. This matches to the most commonly used `nltk.word_tokenize`, minus the sentence tokenizing step.
 
-  (To me it seems like a weird historical thing that NLTK has 2 successive variation on improving the Penn tokenizer, but for now I am matching it and having both.  See [[NLTK#2005]](https://github.com/nltk/nltk/issues/2005).)
+  (To me it seems like a weird historical thing that NLTK has 2 successive variations on improving the Penn tokenizer, but for now, I am matching it and having both.  See [[NLTK#2005]](https://github.com/nltk/nltk/issues/2005).)
 
[…]
