### 2020-01-18 #1956 @leios
I will submit a PR fixing a number of typos in the README and paper. I have a general question on the scope of this software regarding asian language processing, such as Japanese. I know that this is possible with [TinySegmenter.jl](https://github.com/JuliaStrings/TinySegmenter.jl), but is there a way to use this package with WordTokenizers?
