### 2020-01-19 #1956 @Ayushk4
WordTokenizer's default tokeniser doesn't work well on some Asian languages like Chinese, Japanese.

But, we can set up TinySegmenter.jl in the same way as [RevTok.jl example](https://github.com/JuliaText/WordTokenizers.jl#example-setting-tokenizer--revtokjl) given in readme.
Here's an example with the Japanese sample text taken from TinySegmenter.jl.

```
julia> using WordTokenizers

julia> tokenize("आप कैसे हैं?") |> print # Default tokenizer on Hindi
["आप", "कैसे", "हैं", "?"]

julia> tokenize("私の名前は中野です") |> print # Default tokenizer on Japanese
["私の名前は中野です"]

julia> import TinySegmenter

julia> set_tokenizer(TinySegmenter.tokenize)

julia> tokenize("私の名前は中野です") |> print # TinySegmenter's tokenizer on Japanese
SubString{String}["私", "の", "名前", "は", "中野", "です"]
```
