### 2020-01-19 15718a4
Change example setting tokenizer to TinySegmenter.jl's tokenizer
