### 2020-01-19 #1956 @leios
Ah. I was missing the `.tokenize`. I don't know why I didn't put it there, clearly the documentation does the same thing for Revtok. Thanks!
