### 2020-01-19 #1956 @oxinabox
> I have a general question on the scope of this software regarding asian language processing, such as Japanese.

For interest the default tokenizer [TokTok](https://github.com/jonsafari/tok-tok) has:
>  been tested on, and gives reasonably good results for English, Persian, Russian, Czech, French, German, Vietnamese, Tajik, and a few others.

Which is to say a bunch of languages using the Latin (English, French, German, Vietnamese), or the Cyrillic  (Russian, Czech, Tajik), or Persian alphabet (Farsi).
All of which use space as primary token seperation.

We don't yet have anything that supports the morpheme segementation needed for languages like Japanese or Chinese.

I would speculate, just based on the people involved in JuliaText, we are likely to get support for 1 or more Indian languages before anything else.
Most of which, as I understand it are also space seperated.

> I know that this is possible with TinySegmenter.jl, but is there a way to use this package with WordTokenizers?

@Ayushk4 updated the readme now with that example.
Which i think is a much better one,
as RevTok.jl is unmaintained, and also we have our own built in RevTok these days.
