### 2020-01-19 #1956 @leios
I don't see the updated readme yet, if you are going to change it, could you use something like, "ジュリアプログラミング言語が大好きです！"

It translate to "I love the Julia programming language!" but also shows how it segments all three alphabets (it lumps Hiragana words together, which might be an issue for some people).
