### 2020-01-19 41654de
Fix typo in README.md
Co-Authored-By: Lyndon White <oxinabox@ucc.asn.au>
