### 2020-01-20 #1956 @leios
@will-rowe I don't have that option, only "copy link" and "quote reply"
