### 2020-01-20 #1956 @will-rowe
hmmm - did you accept the invite at the top of the thread? Link [here](https://github.com/openjournals/joss-reviews/invitations). If you have already done this, let me know and I'll have to ask someone as I'm out of ideas!
