### 2020-01-29 #1956 @ninjin
Repository:

1. No installation instructions as far as I can see. Although a bit silly for Julia, a single reference to the necessary command is enough.

Paper:

1. I do not believe I have ever seen normalisation as a part of tokenisation, thus I would remove: “Such word tokenization also often includes some normalizing, such as correcting unusual spellings or removing all punctuations.”.

2. It may be worth mentioning that “sentence splitting” is another term often used for “sentence segmentation”. For the record, I have never seen “sentence tokenisation” outside of NLTK, but I am happy to be corrected.

Non-mandatory (just comments, really):

1. It may be wise to consider including purely statistical tokenisation algorithms such as those used by BERT and friends. They are becoming more and more popular and I suspect they may be the standard very soon.

2. For some real competition in regards to performance, see [this](https://github.com/huggingface/tokenizers). It went public post-submission though.

All in all, a nice solid package. Well done!
