### 2020-01-30 2056a9b
Add installation guide
