### 2020-01-30 JuliaText/WordTokenizers.jl#44 @Ayushk4
Add statistical tokenization algorithms
BERT and related models have been using statistical tokenization algorithms. These work well on out-of-vocab words with ML models. High-speed implementations of BPE / WordPiece etc. will be good additions to the package.

Some nice work has been going [here](https://github.com/huggingface/tokenizers)

Thanks to @ ninjin for the suggestion.
