### 2020-01-30 JuliaText/WordTokenizers.jl#44 @oxinabox
this would be great, yes
