### 2020-01-30 85f0fab
Add link to installation section in README
Co-Authored-By: Lyndon White <oxinabox@ucc.asn.au>
