### 2020-02-05 #1956 @oxinabox
> Repository:
>
>     No installation instructions as far as I can see. Although a bit silly for Julia, a single reference to the necessary command is enough.

Done.

> Paper:
>
>     I do not believe I have ever seen normalisation as a part of tokenisation, thus I would remove: “Such word tokenization also often includes some normalizing, such as correcting unusual spellings or removing all punctuations.”.
>

I loosely speaking disagre, but:
I have removed that sentence, as it is overstated. it might be better to say "occucationally" or sometimes. But it doesn't belong right at the start of the paper, but rather in the docstrings of particular tokenizers. Arguably its undersirable behavour.
Examples:
The Twitter Tokenizer (both ours and the original from [NLTK](https://github.com/nltk/nltk/blob/762663e95a1b60affb9d8b9ab40e6a65cec366b8/nltk/tokenize/casual.py#L270-L272
)) normalizes: `waaaaayyyy` into `waaayyy`.

Apparently the [Stanford Tokenizer has a `americanize` flag](https://github.com/nltk/nltk/blob/762663e95a1b60affb9d8b9ab40e6a65cec366b8/nltk/tokenize/stanford.py#L33-L35)
which normalizes: `colour` to `color`.
(Though this really feels like it is stepping out side the bounds of normalizing during tokenization).

On the more minor end, the [original Penn tokenizer](https://github.com/JuliaText/WordTokenizers.jl/blob/master/src/words/penn.sed#L53)  normalizes `"` (double quote) to `''` (pair of single quote).

>    It may be worth mentioning that “sentence splitting” is another term often used for “sentence segmentation”. For the record, I have never seen “sentence tokenisation” outside of NLTK, but I am happy to be corrected.

Done in https://github.com/JuliaText/WordTokenizers.jl/pull/45
my coauthor will post here one merged

> Non-mandatory (just comments, really):
>
>     It may be wise to consider including purely statistical tokenisation algorithms such as those used by BERT and friends. They are becoming more and more popular and I suspect they may be the standard very soon.
>

Good idea, https://github.com/JuliaText/WordTokenizers.jl/issues/44

>     For some real competition in regards to performance, see this. It went public post-submission though.

I think that will indeed be significantly hard to been, but we should see  how we measure up.
https://github.com/JuliaText/WordTokenizers.jl/issues/46
Beyond scope of this submission though.

> All in all, a nice solid package. Well done!

Thanks
