### 2020-02-05 #1956 @ninjin
@oxinabox: Agreed, if it was “occasionally” I would agree. I think there are alternative explanations for your examples though. The first two are examples of canonicalisation that most likely have been integrated into the tokeniser because it is the first piece of code that touches the “raw” text. For PTB, these are necessary for the text to be consistent with the pre-processing that PTB did back in the 90s.

Thanks again for a solid package and I am 200% in support of https://github.com/JuliaText/WordTokenizers.jl/issues/44!
