### 2020-02-06 2ff8b36
Update paper/paper.md
Co-Authored-By: Ayush Kaushal <ayushk4@gmail.com>
