+++ Project.toml
@@ -9,6 +9,8 @@ Unicode = "4ec0a83e-493e-50e2-b9ac-8f72acf5a8f5"
 
 [compat]
 julia = "0.7, 1"
+HTML_Entities= "1"
+StrTables = "1"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
