+++ .travis.yml
@@ -4,9 +4,8 @@ os:
   - linux
   - osx
 julia:
-  - 0.7
   - 1.0
-  - 1.1
+  - 1.3
   - nightly
 notifications:
   email: false
