### 2020-02-07 #1956 @oxinabox
> Thanks @oxinabox . The author names on the archive should match the paper authors please

I am not sure what is controlling the names in the archive.
All the people who as not listed on the paper, who are listed in the archive have made only small corrections to the spelling in the Readme or comments. (Bar one who made a [small fix](https://github.com/JuliaText/WordTokenizers.jl/commit/e3f8d936053f612bd9bf2326b9f35e53fc15c354) to a correr case in the regex.)
These changes are appreciated, but I assume fall below the contribution threashold for co-author ship on a JOSS paper.
Perhaps I am wrong though?
