### 2020-02-07 #1956 @will-rowe
No - you can have the author list as you'd like it (i.e. your in paper). I'm not sure why zenodo has autofilled in all of those authors. You can click edit on the zenodo page and then manually remove the authors which are not on your paper.
