### 2020-02-07 #1956 @danielskatz
You can use a zenodo.json file to tell zenodo what you want automatically, but it's not as easy as it should be be - see https://github.com/zenodo/zenodo/issues/1606 for some discussion on this
