### 2020-02-07 #1956 @will-rowe
Thanks @oxinabox for a nice submission. And many thanks to @leios and @ninjin for your excellent reviews.

Pinging @openjournals/joss-eics to move this to acceptance
