### 2020-02-07 0caaa3d
Update paper.bib
Fixed some capitalization
