+++ paper/paper.bib
@@ -14,7 +14,7 @@
 @Inproceedings{NLTK1,
   Title                    = {NLTK: the natural language toolkit},
   Author                   = {Bird, Steven and Loper, Edward},
-  Booktitle                = {Proceedings of the ACL 2004 on Interactive poster and demonstration sessions},
+  Booktitle                = {Proceedings of the {ACL} 2004 on Interactive poster and demonstration sessions},
   Year                     = {2004},
   Organization             = {Association for Computational Linguistics},
   Pages                    = {31},
@@ -22,7 +22,7 @@
   Url                      = {http://www.aclweb.org/anthology/P04-3031}
 }
 @Book{NLTK2,
-  Title                    = {Natural language processing with Python},
+  Title                    = {Natural language processing with {Python}},
   Author                   = {Bird, Steven and Klein, Ewan and Loper, Edward},
   Publisher                = {O'Reilly Media, Inc.},
   Year                     = {2009},
@@ -42,7 +42,7 @@
 
 @electronic{toktok,
  author = {Dehdari, Jonathan},
- title = {tok-tok: A fast, simple, multilingual tokenizer },
+ title = {tok-tok: A fast, simple, multilingual tokenizer},
  url = {https://github.com/jonsafari/tok-tok},
  urldate = {31.08.2018},
  year = {2015}
