### 2020-02-07 c72345a
Update paper.md
Some grammar and references edits.
