### 2020-02-07 #1956 @kthyng
@oxinabox I've made some changes to your text and bib file in PRs. Please merge what you agree with: [#47](https://github.com/JuliaText/WordTokenizers.jl/pull/47) and [#48](https://github.com/JuliaText/WordTokenizers.jl/pull/48)
