### 2020-02-08 #1956 @oxinabox
> > @oxinabox I'm taking over for the rest of your submission. Can you edit the Zenodo data a little more, so that the title matches that of your paper?
>
> @oxinabox Ok also this part.

I thought I did so.
I don't know why Zenodo is not showing the new title.
