### 2020-02-10 #1956 @kthyng
@oxinabox Is the metadata updated when you look at it online? It still isn't for me. Maybe this will help? https://github.com/geodynamics/best_practices/blob/master/ZenodoBestPractices.md
