### 2020-02-12 #1956 @oxinabox
Ok, not sure what is going on with Zenodo, but i cut a new release on github and that one editted fine.

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3663390.svg)](https://doi.org/10.5281/zenodo.3663390)|

10.5281/zenodo.3663390
