### 2020-02-19 #1956 @kthyng
@openjournals/dev There is something wrong with the paper DOI. This seems to keep happening to me. How can I prevent it/watch for it in the future?
