### 2020-02-19 #1956 @kyleniemeyer
@kthyng the DOI resolves for me, so it may just be a caching issue with your browser? Though sometimes it does take a few minutes or longer for it to initially resolve, which is a Crossref issue, I think.
