### 2020-02-19 #1956 @arfon
> @openjournals/dev There is something wrong with the paper DOI. This seems to keep happening to me. How can I prevent it/watch for it in the future?

It takes a few mins for the DOI to register, plus sometimes, the DOI resolves before the GitHub Pages build is ready (so the PDF doesn't render).

Basically I recommend waiting a few mins before clicking the DOI link, and when rechecking, it sometimes helps to (re)try in incognito/private mode too to help your browser avoid caching issues.
