### 2020-02-24 #1956 @kthyng
@arfon @kyleniemeyer Sorry about the unwarranted alert! When I checked, it was showing up as the orange-ish not-working page, which in the past had meant a different problem, rather than the blue-ish the-DOI-just-hasn't-resolved-yet page.
